requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
Django==5.1.3
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
pillow==11.0.0
pillow-avif-plugin==1.4.6
PySocks==1.7.1
selenium==4.26.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
